refactor(extensions): Refactor how the importing of extensions happens
I agreed with a number of points Mostapha raised in his review.  With the current change, people will be able to see the explicit ImportError if they assign an extension attribute but the module fails to import.  Dealing with the import errors up-front also keeps the rest of the script cleaner and free of try/excepts.