feat(workflow): Add option to reload old results
This change allows users to reload results of a previous study if the Model name and recipe type are the same.  The default is set to False to avoid the (very common) case that a user changes their model geometry but then doesn't see a change in the results as the previous results have been reloaded.

This change also ensures that results from different recipes on the same model do not overwrite one another.  This way, someone could run several different recipes on the same model and re-load all of the results later.